Change and move USB-C connector footprint
The old footprint was causing DRC errors (due to pads being too close to
its holes). Replace it with the one Matthew recommends
(...TYPE-C-31-M-12) and also move it so the edge of the connector
overhangs the board edge to make it easier to put the board in a case or
something.