Add holes for bridge pins on the switches
The switches I intend to use for the project have two extra pins in the
bottom corners, which are simply bridged together. Rather than cut them
off, it seems like a good opportunity to try customizing a footprint, so
create a modified footprint which adds holes for the pins. The Cherry MX
datasheet [0] provided pin locations and recommended hole dimensions.

[0]: https://cdn.sparkfun.com/datasheets/Components/Switches/MX%20Series.pdf